# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Football Service
  description: Football Service Backend API Specification
  contact:
    name: API Support
    email: thscorjs1002@naver.com
  version: 0.0.1
servers:
  - url: 'http://3.34.43.216:8080/api'
    description: production server
tags:
  - name: users
    description: 유저 관련 API
  - name: club
    description: 동호회 관련 API

paths:
  '/users/auth/google':
    get:
      tags:
        - users
      summary: "구글 Oauth 로그인 링크 조회"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  link:
                    example: "https://accounts.google.com/o/oauth2/v2/auth?client_id=436549750929-4pepqro96c8s3sn3a35l84rpeir3brt2.apps.googleusercontent.com&scope=https://www.googleapis.com/auth/userinfo.email%20https://www.googleapis.com/auth/userinfo.profile&response_type=code&access_type=offline&redirect_uri=http%3A%2F%2Flocalhost%3A8080%2Fauth%2Fgoogle"
    post:
      tags:
        - users
      summary: "구글 code로 access token 발급받기"
      parameters:
        - $ref: '#/components/parameters/code'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ..."
                  refresh_token:
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ..."
  '/club':
    post:
      tags:
        - club
      summary: "동호회 개설"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              writeOnly: true
              properties:
                name:
                  $ref: '#/components/parameters/name'
                description:
                  $ref: '#/components/parameters/description'
                main_profile:
                  $ref: '#/components/parameters/main_profile'
                sub_profile:
                  $ref: '#/components/parameters/sub_profile'
                area_code:
                  $ref: '#/components/parameters/area_code'
                cycle_code:
                  $ref: '#/components/parameters/cycle_code'
                gender_code:
                  $ref: '#/components/parameters/gender_code'
                age_group_code:
                  $ref: '#/components/parameters/age_group_code'
      responses:
        201:
          description: Created
        400:
          $ref: '#/components/responses/AlreadyJoinedClubException'
        401:
          $ref: '#/components/responses/CredentialsNotFoundException'
        404:
          $ref: '#/components/responses/CodeNotFoundException'
  '/club/applicant':
    post:
      tags:
        - club
      summary: "동호회원 모집하기"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              writeOnly: true
              properties:
                count:
                  $ref: '#/components/parameters/count'
      responses:
        200:
          description: OK
        401:
          $ref: '#/components/responses/CredentialsNotFoundException'
        403:
          $ref: '#/components/responses/ClubForbiddenException'

components:
  parameters:
    code:
      in: query
      name: code
      schema:
        type: string
        properties:
          code:
            description: "구글에서 받은 code"
            example: "4%2F0AX4XfWiSdMoZzQlu4Rv0P0LieBeFXyp72u321aQWqbFJIbB8nVwdqipXr03o-nbEYiItj1-gw"
    name:
      name: name
      in: path
      required: true
      example: FC동네축구
    description:
      name: description
      in: path
      required: true
      example: 동네축구 동호회입니다 ㅎㅎ
    main_profile:
      name: main_profile
      in: path
      required: true
      example: profile.com
    sub_profile:
      name: sub_profile
      in: path
      required: true
      example: sub-profile.com
    area_code:
      name: area_code
      in: path
      required: true
      example: 1
      description: notion 더미데이터 참고
    cycle_code:
      name: cycle_code
      in: path
      required: true
      example: 1
      description: notion 더미데이터 참고
    gender_code:
      name: gender_code
      in: path
      required: true
      example: 1
      description: notion 더미데이터 참고
    age_group_code:
      name: age_group_code
      in: path
      required: true
      example: 1
      description: notion 더미데이터 참고
    count:
      name: count
      in: path
      required: true
      example: 21
      description: 모집할 인원 수

  responses:
    CredentialsNotFoundException:
      description: 인증 오류
    CodeNotFoundException:
      description: 코드를 찾을 수 없음
    AlreadyJoinedClubException:
      description: 이미 클럽에 가입되어있음
    ClubForbiddenException:
      description: 클럽 관리자가 아님